Rename models/Providers.py to follow snake case convention
