+++ embedchain/models/__init__.py
@@ -1,4 +1,4 @@
 from .EmbeddingFunctions import EmbeddingFunctions  # noqa: F401
-from .Providers import Providers  # noqa: F401
+from .providers import Providers  # noqa: F401
 from .vector_databases import VectorDatabases  # noqa: F401
 from .VectorDimensions import VectorDimensions  # noqa: F401
